% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.biker.r
\name{fit.biker}
\alias{fit.biker}
\title{Plot biker output}
\usage{
fit.biker(obj, data_type = c("data", "hourlyRentals", "stationJourneys",
  "bikeUsage"), fit_type = c("gam", "k.smooth", "loess", "smooth.spline"))
}
\arguments{
\item{obj}{An object of class \code{biker} from \code{\link{loadBikes}}}

\item{data_type}{The type of data to be analysed, either yearly, quarterly, or monthly}

\item{fit_type}{The type of model required, either linear regression (\code{lm}), loess, or smoothing spline (\code{smooth.spline})}
}
\value{
A list of \code{\link[tibble]{tibble}}s which contains hourly, a, and b values for each time series respectively.
}
\description{
Plot biker output
}
\examples{
ans1 = loadBikes('26Jul2017-31Jul2017')
ans2 = fit(ans1)
ans3 = fit(ans1, data_type = 'hourlyRentals', fit_type = 'smooth.spline')
ans4 = fit(ans1, data_type = 'hourlyRentals', fit_type = 'loess')
}
\seealso{
\code{\link{loadBikes}}, \code{\link{plot.biker_fit}}
}
