% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadBikes.r
\name{loadBikes}
\alias{loadBikes}
\title{Load in bike usage data from TFL}
\usage{
loadBikes(range = c("26Jul2017-31Jul2017", "19Jul2017-25Jul2017",
  "12Jul2017-18Jul2017", "05Jul2017-11Jul2017"), ...)
}
\arguments{
\item{range}{A list of weekly date ranges maintained by TFL.}

\item{...}{date ranges corresponding to files on the tfl website}
}
\value{
A list of \code{\link[tibble]{tibble}}s which contains hourly, a, and b values for each time series respectively.
}
\description{
Load in bike usage data from TFL
}
\examples{
data = loadBikes(range = '26Jul2017-31Jul2017')
data = loadBikes(range = '19Jul2017-25Jul2017')
}
\seealso{
\code{\link{loadBikes}}, \code{\link{fit}}
}
