% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.biker.r
\name{map.biker}
\alias{map.biker}
\title{Map biker data}
\usage{
map.biker(obj, data_type = c("stationJourneys", "stationStats"),
  plot_type = c("usage", "rentals"), ...)
}
\arguments{
\item{obj}{Output from the \code{\link{loadBikes}} function}

\item{data_type}{Type of data argument, either stationJourneys (\code{stationJourneys}), stationStats (\code{stationStats})}

\item{plot_type}{Type of data argument, either usage (\code{usage}), rentals (\code{rentals})}
}
\value{
Nothing: just a nice plot
}
\description{
Map biker data
}
\examples{
ans1 = loadBikes('26Jul2017-31Jul2017')
map.biker(obj = ans1, data_type = 'stationStats', 'usage')
map.biker(obj = ans1, data_type = 'stationStats', 'rentals')
}
\seealso{
\code{\link{loadBikes}}
}
