% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bikerr_fit.r
\name{plot.biker_fit}
\alias{plot.biker_fit}
\title{Plot climr output}
\usage{
\method{plot}{biker_fit}(x, time_grid = pretty(x$data$x, n = 100), ...)
}
\arguments{
\item{x}{Output from the \code{\link{fit}} function}

\item{time_grid}{An optional time grid over which to produce fitted values of the model}

\item{...}{Other arguments to plot (not currently implemented)}
}
\value{
Nothing: just a nice plot
}
\description{
Plot climr output
}
\examples{
ans1 = loadBikes('26Jul2017-31Jul2017')
ans2 = fit(ans1)
plot(ans2)
ans3 = fit(ans1, data_type = 'hourlyRentals', fit_type = 'smooth.spline')
plot(ans3)
ans4 = fit(ans1, data_type = 'hourlyRentals', fit_type = 'loess')
plot(ans4)
}
\seealso{
\code{\link{loadBikes}}, \code{\link{fit}}
}
