% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.biker_fit.r
\name{plot}
\alias{plot}
\title{Plot biker output}
\usage{
plot(x, time_grid, ...)
}
\arguments{
\item{x}{Output from the \code{\link{fit}} function}

\item{time_grid}{An optional time grid over which to produce fitted values of the model}

\item{...}{Other arguments to plot (not currently implemented)}
}
\value{
Nothing: just a nice plot
}
\description{
Plot biker output
}
\examples{
ans1 = loadBikes('26Jul2017-31Jul2017')
ans2 = fit(ans1, data_type = 'hourlyRentals', fit_type = 'smooth.spline')
plot(ans2)
ans3 = fit(ans1, data_type = 'hourlyRentals', fit_type = 'loess')
plot(ans3)
ans4 = fit(ans1, data_type = 'dailyRentals', fit_type = 'loess')
plot(ans4)

}
\seealso{
\code{\link{loadBikes}}, \code{\link{fit}}
}
